+++ frontend/src/components/Calendar/CalendarHeader/index.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import moment from 'moment';
 import { useResetRecoilState, useSetRecoilState } from 'recoil';
-import { FaChevronLeft, FaChevronRight, FaCalendarAlt } from 'react-icons/fa';
+
 import { ModalMode, ModalSchedule } from '@stores/calendar';
 import { ColorCode } from '@utils/constants';
-import NewAppointmentBtn from '@components/common/Button';
+import { DateInfoType, weekContentNumber } from '@src/types/calendar';
 
-import { DateInfoType, weekContentNumber } from '../dataStructure';
+import { FaChevronLeft, FaChevronRight, FaCalendarAlt } from 'react-icons/fa';
+import NewAppointmentBtn from '@components/common/Button';
 import { Container, InfoContainer, TodayBtn, ConvertBtn, ConvertBtnContainer, ButtonContainer } from './style';
 
 interface Props {
