+++ frontend/src/components/Calendar/WeeklyCalendar/WeekHeader/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import moment from 'moment';
 import { v4 as uuidv4 } from 'uuid';
-import { DateInfoType, DayCode, weekContentNumber } from '@components/Calendar/dataStructure';
+import { DateInfoType, DayCode, weekContentNumber } from '@src/types/calendar';
 import { Container, DayContainer } from './style';
 
 interface Props {
